refactoring in the naming convention to models
