+++ modules/app/routes/info.js
@@ -1,6 +1,6 @@
 const express = require('express')
 const router = express.Router()
-require('../../../loading/loading')
+require('../../../models')
 require('../../authenticationHelpers')
 const controllers = require('../controllers/info')
