+++ modules/tasks/taskInvite.js
@@ -1,6 +1,6 @@
 const Promise = require('bluebird')
 
-const models = require('../../loading/loading')
+const models = require('../../models')
 const SendMail = require('../mail/mail')
 const i18n = require('i18n')
