+++ modules/tasks/taskSearch.js
@@ -1,5 +1,5 @@
 const Promise = require('bluebird')
-const models = require('../../loading/loading')
+const models = require('../../models')
 
 module.exports = Promise.method(function taskSearch () {
   return models.Task
