+++ test/task.test.js
@@ -4,7 +4,7 @@ const request = require('supertest')
 const expect = require('chai').expect
 const api = require('../server')
 const agent = request.agent(api)
-const models = require('../loading/loading')
+const models = require('../models')
 const { registerAndLogin } = require('./helpers')
 
 describe("tasks", () => {
